Signals, and enemy detection!
The enemy area is linked signal wise to its parent, so that the player can tell the main enemy node to add itself when it's within range, to the main global variable arrays.
NEED TO MAKE THE AREA 2D ITS OWN SCENE SO IT CAN BE ADDED AND CONNECTED AUTOMATICALLY BY THE GOON NODE

Thats for later optimisation.