Enemy attacks!
The enemy will approach the player, and when in range, will pause momentarily, giving the player a chance to attack back, before making their attack. When the player is hit, everyone on both sides is knocked back, and triggers the pause function, which makes them begin to accelerate once more.

I'm surprised this worked this well with so little testing!